Replaced some console logging, reorganized initialization
