+++ css/_navbar.scss
@@ -5,7 +5,7 @@
 	position: relative;
 
 	display: flex;
-	margin-bottom: 30px;
+	margin-bottom: 50px;
 
 	background-image: linear-gradient(#212121, black);
 	box-shadow: 0px 0px 10px 2px black;
