+++ css/style.scss
@@ -1,3 +1,4 @@
 @use "base";
 @use "modal";
 @use "navbar";
+@use "mainpage";
